# aws/request: Minor perf improvement for HandlerList copy, push, and clear