--- a/aws/request/request_internal_test.go
+++ b/aws/request/request_internal_test.go
@@ -2,8 +2,6 @@ package request
 
 import (
 	"testing"
-
-	"github.com/stretchr/testify/assert"
 )
 
 func TestCopy(t *testing.T) {
@@ -15,6 +13,15 @@ func TestCopy(t *testing.T) {
 	req.Handlers = handlers
 
 	r := req.copy()
-	assert.NotEqual(t, req, r)
-	assert.Equal(t, req.Operation.HTTPMethod, r.Operation.HTTPMethod)
+
+	if r == req {
+		t.Fatal("expect request pointer copy to be different")
+	}
+	if r.Operation == req.Operation {
+		t.Errorf("expect request operation pointer to be different")
+	}
+
+	if e, a := req.Operation.HTTPMethod, r.Operation.HTTPMethod; e != a {
+		t.Errorf("expect %q http method, got %q", e, a)
+	}
 }