Minor optimization to the handler list's handling of copying and pushing
request handlers to the handler list.

    benchmark                        old ns/op     new ns/op     delta
    BenchmarkOffsetReader-4          54.7          55.4          +1.28%
    BenchmarkBytesReader-4           0.76          0.74          -2.63%
    BenchmarkNewRequest-4            2629          2682          +2.02%
    BenchmarkHandlersCopy-4          462           366           -20.78%
    BenchmarkHandlersPushBack-4      439           160           -63.55%
    BenchmarkHandlersPushFront-4     633           515           -18.64%
    BenchmarkHandlersClear-4         688           516           -25.00%
    BenchmarkCodegenIterator-4       746562        750308        +0.50%
    BenchmarkEachPageIterator-4      746546        766297        +2.65%

    benchmark                        old allocs     new allocs     delta
    BenchmarkOffsetReader-4          0              0              +0.00%
    BenchmarkBytesReader-4           0              0              +0.00%
    BenchmarkNewRequest-4            22             22             +0.00%
    BenchmarkHandlersCopy-4          4              4              +0.00%
    BenchmarkHandlersPushBack-4      3              1              -66.67%
    BenchmarkHandlersPushFront-4     7              7              +0.00%
    BenchmarkHandlersClear-4         7              7              +0.00%
    BenchmarkCodegenIterator-4       4247           4247           +0.00%
    BenchmarkEachPageIterator-4      4247           4247           +0.00%

    benchmark                        old bytes     new bytes     delta
    BenchmarkOffsetReader-4          3             3             +0.00%
    BenchmarkBytesReader-4           0             0             +0.00%
    BenchmarkNewRequest-4            2688          2688          +0.00%
    BenchmarkHandlersCopy-4          192           192           +0.00%
    BenchmarkHandlersPushBack-4      176           128           -27.27%
    BenchmarkHandlersPushFront-4     352           352           +0.00%
    BenchmarkHandlersClear-4         352           352           +0.00%
    BenchmarkCodegenIterator-4       819700        819723        +0.00%
    BenchmarkEachPageIterator-4      819691        819713        +0.00%